small changes based on feedback

+++ src/backend/distributed/commands/dependencies.c
@@ -42,7 +42,7 @@ static List * GetDependencyCreateDDLCommands(const ObjectAddress *dependency);
 void
 EnsureDependenciesExistsOnAllNodes(const ObjectAddress *target)
 {
-	const uint64 connectionFlag = FORCE_NEW_CONNECTION;
+	const uint32 connectionFlag = FORCE_NEW_CONNECTION;
 
 	/* local variables to work with dependencies */
 	List *dependencies = NIL;
@@ -191,7 +191,7 @@ GetDependencyCreateDDLCommands(const ObjectAddress *dependency)
 void
 ReplicateAllDependenciesToNode(const char *nodeName, int nodePort)
 {
-	const uint64 connectionFlag = FORCE_NEW_CONNECTION;
+	const uint32 connectionFlag = FORCE_NEW_CONNECTION;
 	ListCell *dependencyCell = NULL;
 	List *dependencies = NIL;
 	List *ddlCommands = NIL;
@@ -201,6 +201,13 @@ ReplicateAllDependenciesToNode(const char *nodeName, int nodePort)
 	 * collect all dependencies in creation order and get their ddl commands
 	 */
 	dependencies = GetDistributedObjectAddressList();
+
+	/*
+	 * When dependency lists are getting longer we see a delat in the creation time on the
+	 * workers. We would like to inform the user. Currently we warn for lists greater then
+	 * 100 items, where 100 is an arbitrarily chosen number. If we find it too high or too
+	 * low we can adjust this based on experience.
+	 */
 	if (list_length(dependencies) > 100)
 	{
 		ereport(NOTICE, (errmsg("Replicating postgres objects to node %s:%d", nodeName,
